+++ fiddles/angular2-cli/fiddle-0019-EnterpriseAgGrid/src/app/xyz-grid/xyz-grid.module.ts
@@ -1,0 +1,11 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { XyzGridComponent } from './xyz-grid.component';
+
+@NgModule({
+  declarations: [XyzGridComponent],
+  imports: [
+    CommonModule
+  ]
+})
+export class XyzGridModule { }